test: rewrite core pub/sub tests in fixture-based style
